corrindo erro

+++ status.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 
 class Status(ABC): 
 
-    def __init__(self, nome:str, vida:int, nivel:str) -> None:
+    def __init__(self, nome:str, vida:float, nivel:str) -> None:
         self.nome = nome
         self.vida = vida
         self.vida_base = vida
